io: add cc, fix graph_io compilation
